fixed roof bug, where houses facing -z gitched.
houses face positive z or negative z.

houses facing positive z have their roof blocks placed from

outwards to inwards.

whereas, blocks that are facing the -z had a glitch that placed the roof blocks

from inwards to outwards.

so we create an if statement to check if that block is facing -z

and if it is, create a for loop that creates the roof from outwards to inwards